IsParser changed to work with generic. All parsers and tests are modified according to it.

+++ robe-convert/src/test/java/io/robe/convert/excel/parsers/ParseBigDecimalTest.java
@@ -19,9 +19,9 @@ public class ParseBigDecimalTest {
 
     @Test
     public void parse() throws Exception {
-        BigDecimal actual = (BigDecimal) parseBigDecimal.parse(null, null);
+        BigDecimal actual = parseBigDecimal.parse(null, null);
         assertEquals(null, actual);
-        actual = (BigDecimal) parseBigDecimal.parse("123.12", null);
+        actual = parseBigDecimal.parse("123.12", null);
         assertEquals(expected, actual);
     }
